requirements.txt
matplotlib==3.7.0
numpy==1.23.5
pandas==1.5.2
Pillow==9.5.0
pytest==7.3.1
PyYAML==6.0
scikit_learn==1.2.0
tensorboard==2.10.0
tensorflow==2.10.0
tqdm==4.64.1
